+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Security enhancements
 - Limit for order collection set shipping completed and refactor filter
 
+### Fixed
+- Inventory handling for Magento version >= 2.3 
+
 ### Removed
 - Support for PHP < 7.1
 - Support for Magento < 2.2